store-mme-redis/requirements.txt
functions-framework==3.*
google-cloud-storage
redis
pandas
fsspec
gcsfs
google-cloud-secret-manager
uhashring